refactor: App.tsxのステートとコールバックをuseMailboxカスタムフックに抽出
振る舞いの変更なし。構造的変更のみ。
- useMailbox.tsにステート・useCallback群を集約
- App.tsxはuseMailboxの返り値をコンポーネントに渡すだけに
- tsconfig.jsonからsrc/rendererを除外し、tsconfig.renderer.jsonで管理
