+++ tsconfig.renderer.json
@@ -5,5 +5,5 @@
     "lib": ["ES2020", "DOM"],
     "noEmit": true
   },
-  "include": ["src/renderer/**/*.tsx", "src/renderer/**/*.ts"]
+  "include": ["src/renderer/**/*.ts", "src/renderer/**/*.tsx"]
 }
